create new folders and create files JS
