+++ app.js
@@ -1,6 +1,8 @@
 import { productsBelowTen } from "./apis/belowten.js"
 import { productsOverTen } from "./apis/overten.js"
-import { createTemplate } from "./createTemplates/create-templates.js"
+import createTempleForCartShipping from "./createTemplateFoCartShipping/create-template-for-cart-shipping.js"
+import createTemplate from "./createTemplates/create-templates.js"
+import treatPrices from "./treatPrices/treat-prices.js"
 
 const productsBelow = productsBelowTen['items']
 const productsOver = productsOverTen['items']
@@ -25,48 +27,6 @@ const finalizeBuy = document.querySelector('.finalize')
 
 window.onload = () => allItems.map(item =>
     main.innerHTML += createTemplate(item))
-
-/*FUNÇÃO RESPONSAVEL POR CRIAR O TEMPLATE DOS PRODUTOS */
-
-
-
-/*FUNÇÃO RESPONSAVEL POR CRIAR O TEMPLATE DOS PRODUTOS DENTRO DO CARRINHO */
-
-function createTempleForCartShipping(priceProducts, item) {
-    const { imageUrl, price, ean, name } = priceProducts
-    const priceTreat = treatPrices(price)
-
-    return `
-        <li>
-            <div class="purchase-truffles">
-                <div class=" purchase-image">
-                    <img class="purchase-image-link"
-                        src="${imageUrl}"
-                        alt="">
-                </div>
-                <div class="purchase-info">
-                    <div>
-                        <h1 class="purchase-prod-names">${name}</h1>
-                    </div>
-                    <div class="purchase-cod">
-                        <span>Cód. ${ean}</span>
-                    </div>
-                    <div>
-                        <span class="purchase-price">${priceTreat}</span>
-                    </div>
-                </div>
-                <div class="add-remove-products">
-                    <div>
-                        <span class="material-icons recycle" onclick="removeItem(${item})">
-                            delete_forever
-                        </span>
-                    </div>
-                </div>
-            </div>
-        </li>
-    `
-}
-/* FUNCÇÃO RESPONSAVEL POR GERAR O TEMPLATE DO PREÇO TOTAL NO DOM */
 
 function insertIntoDomPrice(priceTotal) {
     return `
@@ -103,19 +63,6 @@ const insertIntoDomLis = listProducts => {
         ulContainer.innerHTML += createTempleForCartShipping(item, index))
 }
 
-const searchUniqueId = id => allItems.filter(uniqueId =>
-    uniqueId.uniqueId === id
-        ? productsSelecteds.push(uniqueId) : '')
-
-
-const getElementReference = element => {
-    const getParentNode = element.parentNode.parentNode.parentNode
-    const getReferenceTarget = element.tagName === 'BUTTON'
-    const getId = getParentNode.getAttribute('id')
-    getReferenceTarget ? searchUniqueId(getId) : ''
-    verifyAmount()
-}
-
 /*FUNÇÃO QUE CALCULA OS PREÇOS*/
 function calculatePrice(listCartProducts) {
 
@@ -131,111 +78,13 @@ function calculatePrice(listCartProducts) {
 
 /* BLOCO DE FUNÇÕES QUE FAZEM O TRATAMENTO DOS VALORES */
 
-const priceStringfyFortreat = prices => prices.toString().split('').join()
 
-const priceFinalTreatedBelowTen = prices =>
-    priceStringfyFortreat(prices)
-        .replace(',', '.').replace(',', '')
-
-const priceFinalTreatedOverTen = prices =>
-    priceStringfyFortreat(prices)
-        .replace(',', '').replace(',', '.').replace(',', '')
-
-const priceFinalTreatedOverOneHundred = prices =>
-    priceStringfyFortreat(prices)
-        .replace(',', '').replace(',', '').replace(',', '.').replace(',', '')
-
-const priceTreat = price =>
-    Intl.NumberFormat('pt-BR',
-        { style: 'currency', currency: 'BRL' }).format(price)
-
-const treatPrices = prices => {
-
-    const priceToTreated = prices.toString().length
-    const priceLengthThree = priceToTreated <= 3
-    const priceLengthFourOrFive = priceToTreated > 3 && priceToTreated < 5
-    let price;
-
-    if (priceLengthThree) {
-        price = priceFinalTreatedBelowTen(prices)
-    }
-    else if (priceLengthFourOrFive) {
-        price = priceFinalTreatedOverTen(prices)
-    }
-    else {
-        price = priceFinalTreatedOverOneHundred(prices)
-    }
-
-    return priceTreat(price)
-}
-
-
-/* FUNÇÃO QUE VERIFICA SE O CLINTE ATINGIU O MINIMO PARA RECEBER FRETE GRÁTIS */
 
 function checkFreeShipping(amount) {
     amount > 1000
         ? insertMessageFreeShipping()
         : ""
 }
-
-/* OBJETO QUE CONTEM AS FUNÇÕES QUE MANIPULAM OS STYLE NECESSARIOS PARA APLICAÇÃO */
-
-const styles = {
-    styleAmountFlex: () => {
-        styleAmount.style.display = "flex"
-    },
-    styleAmountNone: () => {
-        styleAmount.style.display = "none"
-    },
-    stylePurchaseFlex: () => {
-        purchase.style.display = "flex"
-    },
-    stylePurchaseNone: () => {
-        purchase.style.display = "none"
-    },
-    styleMainFlex: () => {
-        main.style.display = "flex"
-    },
-    styleMainNone: () => {
-        main.style.display = "none"
-    },
-    stylePopUpFlex: () => {
-        popUp.style.display = "flex"
-    },
-    stylePopUpNone: () => {
-        popUp.style.display = "none"
-    },
-    stylePopUpkeepingBuyFlex: () => {
-        popUpkeepingBuy.style.display = "flex"
-    },
-    stylePopUpkeepingBuyNone: () => {
-        popUpkeepingBuy.style.display = "none"
-    },
-    styleModalFlex: () => {
-        modal.style.display = "flex"
-    },
-    styleModalNone: () => {
-        modal.style.display = "none"
-    }
-
-}
-
-/* INSTANCIAMENTO DAS FUNÇÕES QUE MANIPULAM OS ESTILOS */
-
-const popUpBuyFlex = styles['stylePopUpkeepingBuyFlex']
-const popUpBuyNone = styles['stylePopUpkeepingBuyNone']
-const purchaseFlex = styles['stylePurchaseFlex']
-const purchaseNone = styles['stylePurchaseNone']
-const amountFlex = styles['styleAmountFlex']
-const amountNone = styles['styleAmountNone']
-const popUpFlex = styles['stylePopUpFlex']
-const popUpNone = styles['stylePopUpNone']
-const modalFlex = styles['styleModalFlex']
-const modalNone = styles['styleModalNone']
-const mainFlex = styles['styleMainFlex']
-const mainNone = styles['styleMainNone']
-
-/* FUNÇÃO QUE VERIFICA SE O CARRINHO POSSUI ALGUM ITEM E MOSTRA O MODAL */
 
 function verifyItemIntoCart() {
     ulContainer.childElementCount > 0
